Change the way how we count freq - also fix for items deleted at 00:00
Issue was that for freq we counted actual values of "created" and "deleted"
columns, so items "created" at 00:00 and "deleted" at 00:05 were NOT
considered daily (while they should). Also, items "created" at 23:55 and
deleted" at 00:00 were considered "daily" (while they shouldn't).

Two commits ago we fixed it in a dirty way and only for the first case,
this time it's more clean and takes care of second case, also. Now, for
calculating "freq", we take values of "created" and "deleted" columns,
subtract 1 minute from each, and use results for calculations.